final changed for mx block size xfield

+++ src/primitives/block.cpp
@@ -77,10 +77,10 @@ bool CBlockHeader::AbsorbBlockProof(const std::vector<unsigned char>& blockproof
 
 }
 
- uint64_t CBlock::GetHeight() const
- {
+uint32_t CBlock::GetHeight() const
+{
     const CTransaction *coinbase(vtx[0].get());
     if(coinbase->IsCoinBase() && coinbase->vin.size())
         return coinbase->vin[0].prevout.n;
     return 0;
- }
+}
